Move claude-process into backend-cli folder
