+++ src/backend-cli/claude-process.ts
